[Imaging] Add cascaded shadow map support.
The bulk of this change is just plumbing so that GlfSimpleLight can have pointers into multiple entries of GlfSimpleShadowArray.  HdxShadowComputation returns a vector of shadow matrices, instead of just one; the imaging system creates one shadow map per shadow matrix; and the lighting uniforms in GLSL specify a range of applicable shadow maps, instead of just one.

This is a general framework that we can hopefully leverage for lots of cool things (like shadow cubemaps).  The specific cascades implementation resides in two places:
1.) In our shadow matrix computation, in autoframing mode, we now take a parameter for number of cascades (user settable), split the frustum logarithmically, and autoframe each split.
2.) In our lighting code, the "shadowing" function takes a light index (not a shadow index); clips Peye against each shadow frustum; and returns the shadow comparison for the first shadow map in that light's shadow range where the clipping test passes.

(Internal change: 2041853)