Storm volumetric rendering: respecting the grid name and loading the grid from a VDB file identified by the grid name rather than just the first grid.
To achieve this, a texture container are introduced: the texture registry stores a texture container (per VDB file) which stores textures for the different grids in a VDB file.

The texture container is responsible for garbage collecting its own textures (requiring the texture registy to call the new GlfTexture::GarbageCollect()).

(Internal change: 2042134)