Remove redundant image format handling for OpenImageIO
The OIIO plugin had been registered as handling the 'bmp',
'jpg', and 'png' image formats among others. However, the
built-in stb-based image reader also handles these formats
and has a higher precedence than the OIIO plugin. This
means that even if OIIO is installed, it will never be
used to read these image types.

Because of this, we can remove the redundant registrations
from Glf_OIIOImage. This means that users can build USD
against an OIIO library even if it doesn't support these
formats. In particular, this means libjpeg is no longer
a required dependency when using OIIO with USD.

(Internal change: 2042830)