Update PcpChanges API to take const PcpCache*
This more clearly indicates that PcpChanges is
not modifying affected PcpCache objects itself,
but is instead just accumulating changes to be
applied later.

Note that GetCacheChanges still returns a map
of non-const PcpCache pointers for convenience,
since a common pattern might be to apply the
PcpChanges to each PcpCache in the map.

(Internal change: 2042834)