Use PcpCache::FindPrimIndex during prim data composition
PcpCache:ComputePrimIndex is technically not safe to
call from multiple threads as UsdStage was doing in
_ComposeSubtreeImpl. UsdStage relied on computing the
prim indexes ahead of time in _ComposePrimIndexesInParallel
to populate the cache, so that the call to ComputePrimIndex
would just be a map lookup and thus be 'safe'.  However,
in the past this has hidden bugs and led to intermittent
crashes when the necessary prim index hadn't been
pre-computed.

Using FindPrimIndex should avoid these issues and help
expose any lingering bugs in composition.

(Internal change: 2043027)