Skip tests for targets that don't exist in the build
The build will now skip building/registering/installing
tests associated with libraries or plugins that are
not being built due to the particular build configuration.
For example, since plugins are skipped during static
builds, any tests associated with those plugins will
now also be skipped. This avoids errors in cmake due
to pulling on non-existent targets when building these
tests.

This fixes build issues with the new hdTiny example
and test in static builds. In this case, the hdTiny
test was being built even though the plugin was not,
which led to cmake errors.

(Internal change: 2043108)