Fix race in UsdGeomBoundable::ComputeExtentFromPlugins
The _FunctionRegistry singleton was not thread-safe during
initialization. If client code tried to pull on an extent
computation for a type while the registry was still
initializing itself via the call to SubscribeTo, it was
possible that an empty function would be incorrectly
registered for the type. This would cause incorrect bounds
computations for that type for the lifetime of
the process.

This fix ensures the non-registration code paths wait
until initialization is complete.

(Internal change: 2043687)