Applied API schemas now impart their properties as builtins to the prims they are applied to as well as provide fallbacks to typed schemas that include them in their schema definition.
Implementation overview:
When the Usd_PrimData is populated, now, in addition to the type, the "apiSchemas" metadata is composed and provided as part of the type when trying to find or create the Usd_PrimTypeInfo structure. When GetPrimDefinition is called for the first on a Usd_PrimTypeInfo that has additional applied schemas, it will ask the schema to build a composed prim definition from the list of schema types that Usd_PrimTypeInfo will cache and own. When there are no additional applied schemas, the original behavior of returning the pre-populated prim definition directly fromt the schema registry is still used.

As far as the schema registry, it now has the functionality for building a composed prim definition from the existing prim definitions it creates when instantiated. This is used for building composed prim definitions on the fly as mentioned above, but it is also used during instantiation to applied fallback API schemas to the concrete typed prim schemas that have them. This means that schema types that already exist with apiSchemas, e.g. every concrete prim type that derives from UsdLuxLight, will now have additional builtin properties and will return true for HasAPI for its schema defined apiSchemas.

The rest of the changes outside pxr/usd/usd are just fixing of test failures due to the existence of new builtin properties now that applied API schemas provide builtins.

(Internal change: 2044163)