Mark include directories for 3rd-party libraries as system include directories
The INCLUDE_DIRS argument to pxr_library lists include
directories for 3rd-party dependencies. Since USD has
no control over these headers, the build now marks these
as system include directories so that compiler warnings
from these errors are ignored and do not cause strict
builds to fail.

The build had previously treated the headers for Python
this way as a special case, but that behavior is now
covered by this change so the special case can be
removed.

(Internal change: 2046884)