Proper handling for relative paths passed to the HioGlslfx constructor.
Prior to this change, the containingFile parameter for
_ComputeResolvedPath was specified as the current working directory, which
caused that function to improperly resolve relative paths. File-relative
references, those with a ./ or ../ prefix, would fail to resolve when
passed in to the constructor. A relative path that seemed to be resolved
via anchoring would instead resolve via directory search.

Append a trailing slash to the current working directory so that
ArResolver can properly anchor the relative file path to the directory of
the containingFile argument. Change how the resolution test is called,
passing the root-most glslfx file as a file-relative path to exercise this
code path.

(Internal change: 2047672)