[Tf, Usd] Usd.ZipFile.GetFile now returns a bytearray in python
Formerly it returned a string. This was complicated for python 3
support, because the string type changed to always be unicode in the
newer python. A bytearray seems more in keeping with the spirit of
this function, which returns a pointer to a buffer in C++, and is
more portable to python 3.

(Internal change: 2049187)