Fix OpenGL build issue due to bump to CMake 3.12
FindOpenGL changed behavior to prefer GLVND libraries
over the legacy GL library if both are installed starting
with CMake 3.11. Because of this, the version bump to
CMake 3.12 in change 2049628 caused builds to break in
this case because garch currently requires the legacy GL
library (it links explicitly to OPENGL_gl_LIBRARY).

Setting CMP0072 fixes this issue by restoring the
pre-3.11 behavior. We intend to look into garch's
legacy GL dependency as a follow up, which may allow
us to remove this policy flag later.

(Internal change: 2049820)