[usdviewq] Changes to usdview for Python 3 support
This set of changes updates client code to accommodate
functions and syntax that have changed in Python 3.

This set of changes avoids using any compatibility library
in favor of code that works in both Python 2 and 3. This
comes at some performance cost in Python 2 due to creating
full containers instead of using iterators, but we
believe this should not be an issue in practice.

This commit is based off work done by NVIDIA and SideFX.

(Internal change: 2050799)