[build] Update build_usd.py for Python 3 support
build_usd.py will create a USD build with Python 3
if the script is run using Python 3.

When building for Python 3, the script will use
boost 1.70 on Linux and Mac instead of 1.61. This
aligns with VFX Reference Platform CY2020, which
is the first year that Python 3 support is required.
(Note that boost 1.70 is already used for Windows)

This commit is based off work done by NVIDIA and SideFX.

(Internal change: 2051091)