[Hd, UsdImaging] Replace the GetPrimPathForInstanceIndex API with GetScenePrimPath.
GetPrimPathForInstanceIndex returned the instancer of a given rprim, or a null path if there was no instancer.  Most callers of it were expecting it to return the USD prim path to the rprim, and so it was returning the wrong results.

GetScenePrimPath returns the USD prim path of an rprim, as expected by calling code.  It also does the work of turning prims-in-master into instance proxies, which some client code was trying to do itself.

This also fixes up some long-running issues with delegateID-prefixed paths leaking into app code.

To support these changes, the UsdImagingGLEngine picking API has been updated:
- TestIntersection returns a pre-resolved (via GetScenePrimPath) USD prim path, rather than a hydra rprim path.
- GetRprimPathFromPrimId, GetPrimPathFromPrimIdColor, GetPrimPathFromInstanceIndex, and DecodeIDRenderColor have been deleted.
- A new function has been added: DecodeIntersection does the same path resolution as TestIntersection, but given a prim ID/instance ID from an ID render (needed by usdscenemetrics).
- The "element ID" (unsupported) and "worldToLocal" transform (unused) have been deleted.

(Internal change: 2051666)