[Hdx] Always bind the selection buffer when HdxSelectionTask is present.
We were seeing thrashing in some unit tests, and presumably in interactive sessions as well, when you select/deselect objects.  That transition causes "hdxSelectionBuffer", as a bindable buffer, to come in/out of existence, which invalidates all shaders in the scene.

This change always populates "hdxSelectionBuffer" in HdxSelectionTask: when you have the selection task in your rendergraph, your fragment shader always has to query the selection buffer.  This introduces an extra SSBO read per-fragment (though, for a constant location across all fragments), but it should be quick because the selection code early outs if there's no selection.  It reduces the number of shader variations in the hydra ecosystem by half, and means selection/unselection no longer triggers re-batching and shader compilation.

(Internal change: 2055205)