[usdImagingGL] use PYTHON_LIBRARIES instead of PYTHON_LIBRARY in CMakeLists.txt
According to the documentation for FindPythonLibs in CMake 3.12+,
PYTHON_LIBRARY (singular) is intended to be specified by the user and
PYTHON_LIBRARIES (plural) is intended as an output variable. This change
ensures that we use the output variable, and also makes this usage
consistent with pxr/base/tf which also uses PYTHON_LIBRARIES in its
CMakeLists.txt.

(Internal change: 2055949)