Storm texture system: respect the wrap modes authored in a uv texture file.
The HdStUvTextureObject is using Glf to read the wrap modes from the metadata of a uv texture file and translates them into HdWrap. HdStUvSamplerObject resolves the ultimate wrap mode value consulting the sampler parameters authored on a texture node first before falling back to the texture file metadata.

The new texture system replicates the old behavior including that the fallback value for the wrap modes for HwUvTexture_1 vs UsdUvTexture is different and that the obsolete GL_CLAMP is used. The latter required introducing a new HdWrap enum.

HdWrapUseMetadata has been renmed to HdWrapNoOpinion to reflect that this value can be used as return value to indicate that: the texture node has no opinion (meaning we fallback to the metadata in the texture file) or that the metadata in the texture file has no opinion. Aliasing renamed enum values for old clients.

(Internal change: 2057340)