Use the current window's pixel density when determining usdview viewport size
Previously this was using the static method devicePixelRatio() instead of
calling it on the parent window. When testing usdview on Mac, this
caused the viewport size to be incorrect when the window was moved to a
second monitor (since the static method returned the values for the primary
display).

(Internal change: 2069751)