Add support for jump discontinuities in times metadata for value clips.
Jump discontinuities are specified by authoring two entries
in the times metadata with the same stage time but different
clip times. The clip time in the first entry will be used
for time mappings up to the stage time, and the clip time
in the second entry will be used for mappings at and after
the stage time.

Also added more documentation about the stage time to
clip time mapping process and jump discontinuities to
the value clips documentation.

Jump discontinuities are needed to accurately represent
loops for clips -- a discontinuity is needed at the end
of each loop to jump from the last frame of the loop to
the first frame. Previously, the only way to represent
a loop was to author times metadata like:

double2[] times = [(0, 0), (9, 9), (10, 0)]

which caused unwanted time samples at the sub-integer
frames at the end of the loop. Since USD did not know
this was intended to be a loop, it would interpret
the segment between (9, 9) and (10, 0) as reversing
the animation in the clip in those times, which is
not intended. With jump discontinuities, this can now
be encoded as:

double2[] times = [(0, 0), (10, 10), (10, 0)]

(Internal change: 2070275)
(Internal change: 2070302)