[HdEmbree] Fix normal inconsistency with ambient occlusion
There are some inconsistencies in the Embree renderer about the
expectations of normals. This fix makes the code line up with what the
expectation is within other parts of the code when the normal is cached
on hit detection. As noted in the code some possible future work should
be done to make it all consistent:

// there is something odd about how this is used in Embree. Is it
reversed
// here and then when it it used in
//      _ComputeNormal
//      _ComputeColor
// but not when it is used in
//      _EmbreeCullFacess
// this should probably all made to be consistent. What would make most
// sense would be to remove this reversal, and then just change the test
// in _EmbreeCullFaces. this would be the most performant solution.
//

A test was added to make sure this doesn't regress in the future. Note
the threshold isn't one (currently set to 20), in testing this was found
to be slightly noisy which is probably not completely unexpected

(Internal change: 2070432)