Fix Windows junction handling in Tf layer
The Tf layer TfIsDir, TfIsFile and TfIsLink were acting as though any
file with FILE_FLAG_REPARSE_POINT is a symlink.

On Windows, reparse points can have all sorts of purposes, such as
folder junctions (like mounts), or hierarchal storage systems like
OneDrive.

This change ensures that a file with the reparse point flag is actually
a symlink, by attempting to parse its contents, before treating it as
such.

The cost is an extra GetFileAttributes call in the common codepaths of
TfIsDir and TfIsFile when followSymLinks is false (the default), and
attempting to parse the symlink when one is detected.