Change UsdMtlx conversion so that Mtlx "vectorN" datatypes all translate to floatN Sdf datatypes since Sdf vectorNf types impart directionality semantics that mtlx does not, and to "fixed float array" Sdr types, which is how we currently represent role-less float tuples.
In doing so, change conversion tables to indicate that
vectorN <-> floatN is an *exact* type conversion, because
they mean the same thing in the two type systems, and also
improve color2 and color4 translations to use Sdr fixed-arrays
also.

There is a question about how to best handle vectorNarray
types, since Sdr cannot currently represent arrays of tuples,
for which we'll file a separate issue.

(Internal change: 2070909)