Remove transform feedback from non-instanced GPU culling.
Instead, we read the instance count from the (read copy of the) dispatch buffer and write the culled instance count out to the (write) dispatch buffer.

By not using transform feedback the code path is better supported on Metal and Vulkan where transform feeback is not available as a core feature.

Some documentation fixes.

(Internal change: 2074818)